major rewrite/overhaul that now allows for hardware profiles.  created initially to address sda vs vda disk devices, this will easily be adapted to hp proliant funky disk names, etc...  sit back and get nautious as you inspect the code.
AND, you can now run EVERYTHING from the libvirt host.  No need to install ansible on the bastion, clone the repo, etc...

AND, the libvirt host installs the ssh keys on bastion, so password is also no longer required in the ks.cfg.  I'll remove the default after a little more testing

The ENTIRE process for a libvirt deployment will be completely automated very soon.